+++ templates/react-gun-dapp/src/app.js
@@ -1,0 +1,32 @@
+import React from 'react';
+import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
+
+import './app.css';
+import Home from './pages/home';
+import About from './pages/about';
+
+function App() {
+  return (
+    <div className="app">
+      <Routes>
+        <Route exact path="/" element={<Home />} />
+        <Route path="/about" element={<About />} />
+        <Route path="/home" element={<Home />} />
+        <Route path="*" element={<Navigate to="/" />} />
+      </Routes>
+    </div>
+  );
+}
+
+const WrappedApp = App;
+
+export default () => {
+  // While the blocklet is deploy to a sub path, this will be work properly.
+  const basename = window?.blocklet?.prefix || '/';
+
+  return (
+    <Router basename={basename}>
+      <WrappedApp />
+    </Router>
+  );
+};