+++ templates/react-static/src/pages/home.js
@@ -1,11 +1,15 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+
 import logo from '../logo.svg';
 
 const Home = () => {
   return (
     <header className="app-header">
       <img src={logo} className="app-logo" alt="logo" />
+      <pre style={{ textAlign: 'left' }}>
+        <code>window.blocklet = {JSON.stringify(window.blocklet, null, 2)}</code>
+      </pre>
       <Link className="app-link" to="/about">
         About
       </Link>