+++ test/blocklet-server.test.js
@@ -8,11 +8,10 @@ test('blocklet server is installed', async () => {
 });
 
 test(
-  'blocklet server version is 1.7.0',
+  'blocklet server version is bigger than 1.7.0',
   async () => {
     const version = await getServerVersion();
-    expect(version).toBe('1.7.0');
-    expect(semver.satisfies(version, '>= 1.7.0')).toBe(true);
+    expect(semver.satisfies(version, '>=1.7.0')).toBe(true);
   },
   -1
 );
@@ -31,7 +30,6 @@ test(
   'blocklet server directory',
   async () => {
     const directory = await getServerDirectory();
-    console.log({ directory });
     expect(directory).toBe('/Users/han/workspace/an-test/prod');
   },
   -1